task 5

+++ python_variable_annotations/5-sum_list.py
@@ -2,5 +2,5 @@
 """Takes a list of floats and returns their sum"""
 
 
-def sum_list(lista: list[float]) -> int:
-	return int(sum(lista))
+def sum_list(lista: list[float]) -> float:
+	return sum(lista)
